// src/app/interfaces/diasventausuario-response.ts
export interface DiasVentaUsuarioInterface {
    IDVENTA: number;
    IDMOV:   number;
    LUN:     string | boolean;
    MAR:     string | boolean;
    MIE:     string | boolean;
    JUE:     string | boolean;
    VIE:     string | boolean;
    SAB:     string | boolean;
    DOM:     string | boolean;
}